adjusted

+++ src/Pages/CreateFood/index.jsx
@@ -29,6 +29,7 @@ export function CreateFood({ ...rest }) {
   const [ingredients, setIngredients] = useState([])
   const [newIngredient, setNewIngredient] = useState('')
   const [foodImage, setFoodImage] = useState(null)
+  const [foodImageFile, setFoodImageFile] = useState(null)
   const inputFileRef = useRef(null)
   const navigate = useNavigate()
 
@@ -36,29 +37,19 @@ export function CreateFood({ ...rest }) {
     async function fetchProductData() {
       if (id) {
         try {
-          console.log(`Tentando buscar produto com ID: ${id}`); // Log para debug
           const response = await api.get(`/product/${id}`)
           const { name, description, price, categories_id, ingredients, image } = response.data
-          
+
           setName(name)
           setDescription(description)
           setPrice(price)
           setCategory(categories_id)
           setIngredients(ingredients)
-          setFoodImage(image)
+          setFoodImage(`${import.meta.env.VITE_API_URL}/files/${image}`)
         } catch (error) {
-          console.error('Erro ao carregar os dados do produto:', error)
-          if (error.response) {
-            console.error('Resposta do servidor:', error.response.data)
-            console.error('Status do erro:', error.response.status)
-          }
-          if (error.response && error.response.status === 404) {
-            alert(`Produto com ID ${id} não encontrado. Verifique se o ID está correto.`)
-          } else {
-            alert('Erro ao carregar os dados do produto. Por favor, tente novamente mais tarde.')
-          }
-          // Opcionalmente, redirecione para uma página segura
-          // navigate('/admin')
+          navigate('/')
+
+          alert('Erro ao carregar os dados do produto. Por favor, tente novamente mais tarde.')
         }
       }
     }
@@ -67,7 +58,7 @@ export function CreateFood({ ...rest }) {
   }, [id, navigate])
 
   function handleBack() {
-    navigate('/admin')
+    navigate('/')
   }
 
   function handleAddIngredient() {
@@ -83,54 +74,13 @@ export function CreateFood({ ...rest }) {
     )
   }
 
-  async function handleChangeImageFood(e) {
-    const file = e.target.files[0];
-    if (file) {
-      try {
-        const formData = new FormData();
-        formData.append('image', file);
-  
-        // Fazendo o upload da imagem
-        const response = await api.post('/upload-image', formData, {
-          headers: {
-            'Content-Type': 'multipart/form-data',
-          },
-        });
-  
-        if (response.data && response.data.imageUrl) {
-          const imageUrl = response.data.imageUrl;
-  
-          // Atualizando a URL da imagem no banco de dados
-          const updateResponse = await api.put('/update-food-image', {
-            imageUrl, // URL da imagem
-            foodId: id, // Verifique se "id" está definido corretamente
-          });
-  
-          if (updateResponse.status === 200) {
-            setFoodImage(imageUrl); // Atualiza o estado da imagem localmente
-          } else {
-            throw new Error('Falha ao atualizar imagem no banco de dados');
-          }
-        } else {
-          throw new Error('Falha ao obter URL da imagem');
-        }
-      } catch (error) {
-        console.error('Erro ao fazer upload da imagem:', error);
-        if (error.response) {
-          console.error('Status do erro:', error.response.status);
-          console.error('Resposta do servidor:', error.response.data);
-          if (error.response.status === 404) {
-            alert('Erro 404: Endpoint de upload não encontrado. Verifique a rota de upload.');
-          } else {
-            alert(`Erro ${error.response.status}: ${error.response.data.message || 'Erro desconhecido'}`);
-          }
-        } else {
-          alert('Erro ao fazer upload da imagem. Por favor, tente novamente.');
-        }
-      }
-    }
+  function handleChangeImageFood(e) {
+    const file = e.target.files[0]
+    setFoodImageFile(file)
+
+    const imagePreview = URL.createObjectURL(file)
+    setFoodImage(imagePreview)
   }
-  
 
   function handleClickUpload() {
     inputFileRef.current.click()
@@ -138,63 +88,50 @@ export function CreateFood({ ...rest }) {
 
   async function handleNewProduct() {
     try {
-      if (!name || !description || ingredients.length === 0 || !foodImage || !price || !categories_id) {
+      if (!validateRequiredFields()) {
         alert('Por favor, preencha todos os campos obrigatórios.')
         return
       }
 
-      // Primeiro, fazemos o upload da imagem
-      const imageFormData = new FormData()
-      imageFormData.append('image', foodImage)
-
-      const imageUploadResponse = await api.post('/upload-image', imageFormData, {
-        headers: {
-          'Content-Type': 'multipart/form-data'
-        }
-      })
-
-      if (!imageUploadResponse.data || !imageUploadResponse.data.imageUrl) {
-        throw new Error('Falha ao obter URL da imagem após o upload')
-      }
-
-      const imageUrl = imageUploadResponse.data.imageUrl
-
-      // Agora, criamos o produto com a URL da imagem
       const productFormData = new FormData()
       productFormData.append('name', name)
       productFormData.append('description', description)
       ingredients.forEach((ingredient, index) => {
         productFormData.append(`ingredients[${index}]`, ingredient)
       })
-      productFormData.append('image', imageUrl)
+  
+      if (foodImageFile instanceof File) {
+        productFormData.append('image', foodImageFile)
+      } else {
+        console.warn('Aviso: Arquivo de imagem inválido ou não fornecido.')
+      }
+  
       productFormData.append('price', price)
       productFormData.append('categories_id', categories_id)
 
+      console.log('FormData:', productFormData)
       await api.post('/product', productFormData, {
         headers: {
           'Content-Type': 'multipart/form-data'
         }
       })
 
-      alert('Novo prato cadastrado!')
-      navigate('/admin')
+      alert('Novo prato cadastrado com sucesso!')
+      navigate('/')
     } catch (error) {
-      console.error('Erro ao criar o produto:', error)
-      if (error.response) {
-        console.error('Status do erro:', error.response.status)
-        console.error('Resposta do servidor:', error.response.data)
-        alert(`Erro ${error.response.status}: ${error.response.data.message || 'Erro desconhecido'}`)
-      } else if (error.message) {
-        alert(error.message)
-      } else {
-        alert('Erro ao criar o produto. Tente novamente.')
-      }
+      console.error('Erro ao criar o novo produto:', error)
+      alert('Erro ao criar o novo produto. Por favor, tente novamente mais tarde.')
     }
   }
 
+  function validateRequiredFields() {
+    return name && description && ingredients.length > 0 && foodImage && price && categories_id
+  }
+
+
   async function handleUpdatedProduct() {
     try {
-      if (!name || !description || !price ) {
+      if (!name || !description || !price) {
         alert('Por favor, preencha todos os campos obrigatórios.')
         return
       }
@@ -205,58 +142,41 @@ export function CreateFood({ ...rest }) {
       ingredients.forEach((ingredient, index) => {
         formData.append(`ingredients[${index}]`, ingredient)
       })
-      if (foodImage instanceof File) {
-        formData.append('image', foodImage)
+
+      if (foodImageFile instanceof File) {
+        formData.append('image', foodImageFile)
       }
+
       formData.append('price', price)
       formData.append('categories_id', categories_id)
 
-      const response = await api.put(`/product/${id}`, formData, {
+      await api.put(`/product/${id}`, formData, {
         headers: {
           'Content-Type': 'multipart/form-data'
         }
       })
 
-      console.log('Resposta do servidor:', response.data)
       alert('Produto atualizado com sucesso!')
-      navigate('/admin')
+      navigate('/')
     } catch (error) {
       console.error('Erro ao atualizar o produto:', error)
-      if (error.response) {
-        console.error('Resposta do servidor:', error.response.data)
-        console.error('Status do erro:', error.response.status)
-        alert(`Erro ao atualizar o produto: ${error.response.data.message || 'Erro desconhecido'}`)
-      } else if (error.request) {
-        console.error('Nenhuma resposta recebida:', error.request)
-        alert('Erro de conexão. Por favor, verifique sua internet e tente novamente.')
-      } else {
-        console.error('Erro ao configurar a requisição:', error.message)
-        alert('Erro ao processar a solicitação. Por favor, tente novamente.')
-      }
+      alert('Erro ao atualizar o produto. Por favor, tente novamente mais tarde.')
     }
   }
 
   async function handleDeleteProduct() {
     try {
-      const response = await api.delete(`/product/${id}`)
-      console.log('Resposta do servidor:', response.data)
+      const confirmDelete = window.confirm('Tem certeza que deseja excluir este produto?');
+      if (!confirmDelete) {
+        return; // Se o usuário não confirmar, interrompe a execução da função
+      }
+      await api.delete(`/product/${id}`)
       alert('Produto excluído com sucesso!')
-      navigate('/admin')
+      navigate('/')
     } catch (error) {
       console.error('Erro ao excluir o produto:', error)
-      if (error.response) {
-        console.error('Resposta do servidor:', error.response.data)
-        console.error('Status do erro:', error.response.status)
-        alert(`Erro ao excluir o produto: ${error.response.data.message || 'Erro desconhecido'}`)
-      } else if (error.request) {
-        console.error('Nenhuma resposta recebida:', error.request)
-        alert('Erro de conexão. Por favor, verifique sua internet e tente novamente.')
-      } else {
-        console.error('Erro ao configurar a requisição:', error.message)
-        alert('Erro ao excluir o produto. Tente novamente.')
-      }
+      alert('Erro ao excluir o produto. Por favor, tente novamente mais tarde.')
     }
-
   }
 
   return (
@@ -270,6 +190,15 @@ export function CreateFood({ ...rest }) {
             title="Voltar"
           />
           <h1>{id ? 'Editar prato' : 'Novo prato'}</h1>
+          {foodImage && (
+            <div>
+              <img
+                src={foodImage}
+                alt="Pré-visualização da comida"
+                style={{ maxWidth: '40px', width: '50%', maxHeight: '20px', marginTop: '5px' }}
+              />
+            </div>
+          )}
         </GoBack>
 
         <CardDetails>
@@ -281,22 +210,14 @@ export function CreateFood({ ...rest }) {
               title="Selecione a imagem"
               onClick={handleClickUpload}
             />
+
             <input
               type="file"
-              accept="image/*"
+              id="image-food"
               ref={inputFileRef}
               style={{ display: 'none' }}
               onChange={handleChangeImageFood}
             />
-            {foodImage && (
-              <div>
-                <img
-                  src={typeof foodImage === 'string' ? foodImage : URL.createObjectURL(foodImage)}
-                  alt="Pré-visualização da comida"
-                  style={{ maxWidth: '40px', width: '50%', maxHeight: '20px', marginTop: '5px' }}
-                />
-              </div>
-            )}
           </ButtonUpload>
 
           <Input
@@ -362,6 +283,7 @@ export function CreateFood({ ...rest }) {
 
         <SubmitButton>
           <Button
+            type="submit"
             className="Submit"
             title={id ? 'Salvar alterações' : 'Criar novo prato'}
             onClick={id ? handleUpdatedProduct : handleNewProduct}
